# api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Register a customer or user account",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Account created; email verification is required before login"
          },
          "400": {
            "description": "Validation failed"
          },
          "409": {
            "description": "Email or phone is already registered"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App",
        "security": []
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Log in with account credentials",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AuthCredentialsDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Authentication successful"
          },
          "401": {
            "description": "Invalid credentials or email has not been verified"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App",
        "security": []
      }
    },
    "/api/v1/auth/oauth/google": {
      "post": {
        "operationId": "AuthController_googleLogin",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Sign in or register with a Google identity token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Google authentication successful"
          },
          "401": {
            "description": "Invalid Google identity token"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App",
        "security": []
      }
    },
    "/api/v1/auth/oauth/apple": {
      "post": {
        "operationId": "AuthController_appleLogin",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Sign in or register with an Apple identity token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SocialLoginDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Apple authentication successful"
          },
          "401": {
            "description": "Invalid Apple identity token"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App",
        "security": []
      }
    },
    "/api/v1/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Exchange a refresh token for a new access token",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Access token refreshed"
          },
          "401": {
            "description": "Refresh token is invalid or expired"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App",
        "security": []
      }
    },
    "/api/v1/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "summary": "[CUSTOMER APP] Revoke a refresh token and log out",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefreshTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged out successfully"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/auth/forgot-password": {
      "post": {
        "operationId": "AuthController_requestPasswordReset",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Request a password-reset flow",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password-reset request accepted"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App",
        "security": []
      }
    },
    "/api/v1/auth/reset-password": {
      "post": {
        "operationId": "AuthController_resetPassword",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Reset password using only the emailed code",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Password reset successfully"
          },
          "401": {
            "description": "Reset code is invalid, expired, or attempt-limited"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App",
        "security": []
      }
    },
    "/api/v1/auth/verify-email": {
      "post": {
        "operationId": "AuthController_verifyEmail",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Verify an email using the emailed 6-digit code",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VerifyEmailCodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Email verified"
          },
          "401": {
            "description": "Code invalid or expired"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App",
        "security": []
      }
    },
    "/api/v1/auth/request-email-verification": {
      "post": {
        "operationId": "AuthController_requestEmailVerification",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Send a new 6-digit email verification code",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification request accepted"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App",
        "security": []
      }
    },
    "/api/v1/auth/resend-verification": {
      "post": {
        "operationId": "AuthController_resendVerification",
        "x-audience": "Public / Customer App",
        "summary": "[PUBLIC / CUSTOMER APP] Resend a 6-digit email verification code",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PasswordResetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Verification request accepted"
          }
        },
        "tags": [
          "Customer App - Authentication"
        ],
        "description": "Audience: Public / Customer App",
        "x-audience-label": "Public / Customer App",
        "security": []
      }
    },
...
    "/api/v1/customers/me/addresses": {
      "get": {
        "operationId": "AddressesController_list",
        "summary": "[CUSTOMER APP] List",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Addresses"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      },
      "post": {
        "operationId": "AddressesController_create",
        "summary": "[CUSTOMER APP] Create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateAddressDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Addresses"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
    "/api/v1/customers/me/addresses/{id}": {
      "patch": {
        "operationId": "AddressesController_update",
        "summary": "[CUSTOMER APP] Update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAddressDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Addresses"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      },
      "delete": {
        "operationId": "AddressesController_remove",
        "summary": "[CUSTOMER APP] Remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response"
          },
          "400": {
            "description": "Validation or business-rule error"
          },
          "401": {
            "description": "Unauthorized"
          }
        },
        "tags": [
          "Customer App - Addresses"
        ],
        "description": "Audience: Customer App",
        "x-audience-label": "Customer App"
      }
    },
...
  },
  "info": {
    "title": "La Favola Pizza Restaurant API",
    "description": "Routes are grouped by audience. Customer app developers should use sections prefixed \"Customer App\" and \"Audience: Public / Customer App\". Administrative and staff operations declare their required roles in their Audience tag and x-required-roles metadata.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "RegisterDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "Email address of the user",
            "example": "john.doe@example.com",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "description": "Password of the user. Must contain at least one uppercase letter, one lowercase letter, and one number.",
            "example": "StrongPass123",
            "minLength": 8,
            "maxLength": 128
          },
          "fullName": {
            "type": "string",
            "description": "Full name of the user (only letters and spaces allowed)",
            "example": "John Doe",
            "minLength": 2,
            "maxLength": 100
          },
          "phone": {
            "type": "string",
            "example": "+393331234567"
          }
        },
        "required": [
          "email",
          "password",
          "fullName"
        ]
      },
      "AuthCredentialsDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address used for login",
            "example": "jane.doe@example.com",
            "maxLength": 255
          },
          "password": {
            "type": "string",
            "description": "Password of the user",
            "example": "StrongPass123!",
            "minLength": 1,
            "maxLength": 255
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "SocialLoginDto": {
        "type": "object",
        "properties": {
          "idToken": {
            "type": "string",
            "description": "The identity token returned by Google or Apple",
            "example": "eyJhbGciOiJSUzI1NiIsImtpZCI6Ii4uLiJ9..."
          },
          "fullName": {
            "type": "string",
            "description": "Name supplied on first sign-in (especially for Apple, which only supplies it once)",
            "example": "Mario Rossi"
          }
        },
        "required": [
          "idToken"
        ]
      },
      "RefreshTokenDto": {
        "type": "object",
        "properties": {}
      },
      "PasswordResetDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "customer@example.com"
          }
        },
        "required": [
          "email"
        ]
      },
      "ResetPasswordDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "123456",
            "pattern": "^\\d{6}$",
            "minLength": 6,
            "maxLength": 6,
            "description": "Single-use code sent by email"
          },
          "password": {
            "type": "string",
            "example": "NewSecurePass1",
            "minLength": 8,
            "maxLength": 128
          }
        },
        "required": [
          "code",
          "password"
        ]
      },
      "VerifyEmailCodeDto": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "123456",
            "pattern": "^\\d{6}$",
            "minLength": 6,
            "maxLength": 6,
            "description": "Single-use email verification code"
          }
        },
        "required": [
          "code"
        ]
      },
...
      "CreateAddressDto": {
        "type": "object",
        "properties": {}
      },
      "UpdateAddressDto": {
        "type": "object",
        "properties": {}
      },
...
    }
  },
  "security": [
    {
      "JWT-auth": []
    }
  ]
}